T#803 P3-H: Extract teams module from server.ts
Sub-task H of T#786 Phase 3. Extract 10 teams handlers (~162 LoC) + 3 helper
functions (validateTeamName, sanitizeInput, beastExists) from server.ts into
src/teams/routes.ts.

## Changes

- New: src/teams/routes.ts (177L) — registerTeamsRoutes(app, sqlite, helpers)
  with TeamsHelpers DI (hasSessionAuth); 3 closure helpers + 10 handlers
  extracted verbatim
- server.ts: 3919L → 3757L (-162L)

## Handlers
- GET /api/teams
- POST /api/teams
- GET /api/teams/:id
- PATCH /api/teams/:id
- POST /api/teams/:id/members
- DELETE /api/teams/:id/members/:beast
- POST /api/teams/:id/projects
- DELETE /api/teams/:id/projects/:projectId
- DELETE /api/teams/:id (creator-or-gorn gate preserved verbatim)
- GET /api/teams/beast/:beast

## Test plan
- [x] bun build clean (Norm #84 Layer 1)
- [x] Module registration before SPA fallback
- [ ] Post-deploy smoke: GET /api/teams → 200 + {teams, total}

Reviewer: @zaghnal per task assignment.
